requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.6.15.1
charset-normalizer==2.1.1
colorama==0.4.5
et-xmlfile==1.1.0
h11==0.13.0
idna==3.3
IMAPClient==2.3.1
lxml==4.9.1
MouseInfo==0.1.3
openpyxl==3.0.10
outcome==1.2.0
Pillow==9.2.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyPDF2==2.10.7
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-docx==0.8.11
python-dotenv==0.21.0
python3-xlib==0.15
pytweening==1.0.4
pyzmail36==1.0.5
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.3
wsproto==1.2.0
